Add FcPreprocessFilename speeding up FcCompareFilename.
FcPreprocessFilename calculates hash of the filename to fast reject not matching filenames. Filenames are compared both case sensitive and case insensitive. The hash is calculated for case insensitive variant. That way if hash is not matching, we know for sure that both kinds of string comparison would fail.

In adition FcPreprocessFilename remembers whether the filename contains any globs ('?', '*') so that FcCompareFilename cam quickly skip over FcStrGlobMatch in case there are no globs.

TODO: Add benchmark numbers.