Reduce heap allocations when iterating FcStrSet.
To iterate FcStrSet was has to create FcStrList (iterator) using FcStrListCreate function and destroy it at the end using FcStrListDone function. FcStrList is allocated on heap, however in many cases it is needed only in the scope of one function and could be easily on stack.

This commit splits the (de-)initialization of FcStrList away from (de)allocation and introduces FcStrListInitialize and FcStrListRelease functions. Then it converts all uses of FcStrListCreate and FcStrListDone to them where possible.
This also lets us remove some fail paths because FcStrListInitialize unlike FcStrListCreate can not fail.