Optimize FcStrCaseWalkerNext.
FcStrCaseWalkerNext is in several hot paths, so it is worth optimizing it even if it means having some duplicate code.

At this moment it is used in 3 ways:
* Without any delimiter (often)
* With a single character " " (often)
* With two character delimiters " -" (only from FcComparePostScript)

It helps to make separate function for each of those cases.
Using strchr in the multiple delimiter case is overkill because the string of delimiters is normally very short (currently always 2 characters). Any clever optimizations that strchr may do are not worth it, naive inline implementation is faster.

TODO: Benchmark data.