Correct score selection in FcFontSetMatchInternal
FcFontSetMatchInternal was using the strength of the first element in
deciding which score to use. This is wrong as we should look at the
priority instead.

Instead of building two scores to choose from, we now only use one
score variable and decide whether it should be used for strong, weak,
or both by testing for matcher->{strong,weak} == priority. Doing so
ensures perfect emulation of FcCompare() behavior for each priority.

Fixes michalsrb/fontconfig#2.

A previous version of this commit decided between a strong and a weak
score by testing for matcher->strong == priority. This fix is
incomplete: for priorities that share the same value for strong and
weak, FcCompareValueList would use the smaller of those two when called
by FcCompare as two pointers to "score" coincide; the previous fix does
not similate this feature.